Fix Thunderstore API to allow bigger file sizes. (Thanks Valheim modders for giant fucking mods!)

+++ AssemblyPublicizer/AssemblyPublicizer.cs
@@ -406,7 +406,7 @@ namespace CabbageCrow.AssemblyPublicizer
             public string? website_url { get; set; }
             public bool is_active { get; set; }
             public string? uuid4 { get; set; }
-            public int file_size { get; set; }
+            public float file_size { get; set; }
         }
 
         public class PackageListing
